getQuery, checkNumbers functions moved to helpers folder, selected styles in header, searchBar rendering, some little issues fixed
