+++ fe/src/components/Header/Header.scss
@@ -77,7 +77,7 @@ html {
     height: 14px;
 
     font-family: 'Mont';
-    left: 26%;
+    left: 25%;
     top: 28%;
     font-weight: 600;
     font-size: 9px;
@@ -88,6 +88,16 @@ html {
     color: #fff;
     background-color: #4219d0;
     text-decoration: none;
+
+    @include onTablet {
+      left: 55%;
+      top: 30%;
+    }
+
+    @include onDesktop {
+      left: 50%;
+      top: 28%;
+    }
   }
 
   &__cart {
@@ -97,7 +107,7 @@ html {
     height: 14px;
 
     font-family: 'Mont';
-    left: 76%;
+    left: 75%;
     top: 30%;
     font-weight: 600;
     font-size: 9px;
@@ -108,6 +118,16 @@ html {
     color: #fff;
     background-color: #4219d0;
     text-decoration: none;
+
+    @include onTablet {
+      left: 55%;
+      top: 30%;
+    }
+
+    @include onDesktop {
+      left: 50%;
+      top: 30%;
+    }
   }
 
   &__menu {
@@ -124,4 +144,47 @@ html {
   width: 15px;
   height: 15px;
 }
+
+.favorites-icon {
+  position: relative;
+
+  &:hover {
+    &::after {
+      content: '';
+      position: absolute;
+      bottom: 0;
+      left: 0;
+      right: 0;
+      height: 4px;
+      background-color: #0f0f11;
+    }
+  }
+}
+
+.cart-icon {
+  position: relative;
+
+  &:hover {
+    &::after {
+      content: '';
+      position: absolute;
+      bottom: 0;
+      left: 0;
+      right: 0;
+      height: 4px;
+      background-color: #0f0f11;
+    }
+  }
+}
+
+.is-active::after {
+  content: '';
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  height: 4px;
+  background-color: #0f0f11;
+}
+
