+++ fe/src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { Link } from 'react-router-dom';
+import { NavLink } from 'react-router-dom';
 import StateContext from '../../components/Context/Context';
 
 import { BurgerMenu } from '../BurgerMenu';
@@ -11,6 +11,7 @@ import './Header.scss';
 import heart from '../../assets/images/icons/heart.svg';
 import bag from '../../assets/images/icons/bag.svg';
 import union from '../../assets/images/icons/union.svg';
+import classNames from 'classnames';
 
 type Props = {
   showBurger: boolean;
@@ -31,23 +32,37 @@ export const Header: React.FC<Props> = ({ showBurger, setShowBurger }) => {
         <Navigation />
         <BurgerMenu showBurger={showBurger} setShowBurger={setShowBurger} />
         <div className="header__icon">
-          <Link to="/favourites">
-            <div className="icon-box">
-              <img src={heart} alt="like" className="icon" />
-              {!!favouriteItems.length
-                && <div className="icon-box__favourite">
-                  {favouriteItems.length}
+          <div className="favorites-icon">
+            <NavLink
+              to="/favourites"
+              className={({ isActive }) =>
+                classNames({ 'is-active': isActive })
+              }
+            >
+              <div className="icon-box">
+                <img src={heart} alt="like" className="icon" />
+                {!!favouriteItems.length
+                  && <div className="icon-box__favourite">
+                    {favouriteItems.length}
+                  </div>}
+              </div>
+            </NavLink>
+          </div>
+          <div className="cart-icon">
+            <NavLink
+              to="/cart"
+              className={({ isActive }) =>
+                classNames({ 'is-active': isActive })
+              }
+            >
+              <div className="icon-box">
+                <img src={bag} alt="bag" className="icon" />
+                {!!cartItems.length && <div className="icon-box__cart">
+                  {cartItems.length}
                 </div>}
-            </div>
-          </Link>
-          <Link to="/cart">
-            <div className="icon-box">
-              <img src={bag} alt="bag" className="icon" />
-              {!!cartItems.length && <div className="icon-box__cart">
-                {cartItems.length}
-              </div>}
-            </div>
-          </Link>
+              </div>
+            </NavLink>
+          </div>
           <div
             className="icon-box icon-box__menu"
             onClick={() => changeBurgerStatus(showBurger)}
