+++ fe/src/components/PageNavLink/PageNavLink.scss
@@ -27,6 +27,16 @@
 
   &:hover {
     color: #0f0f11;
+
+    &::after {
+      content: '';
+      position: absolute;
+      bottom: 0;
+      left: 0;
+      right: 0;
+      height: 4px;
+      background-color: #0f0f11;
+    }
   }
 }
