+++ fe/src/components/Pagination/Pagination.tsx
@@ -1,15 +1,6 @@
 import React from 'react';
 import classNames from 'classnames';
-
-export function getNumbers(from: number, to: number): number[] {
-  const numbers = [] as number[];
-
-  for (let n = from; n <= to; n += 1) {
-    numbers.push(n);
-  }
-
-  return numbers;
-}
+import { getNumbers } from '../../helpers/getNumbers';
 
 type Props = {
   totalPages: number;
