+++ fe/src/helpers/checkQueryIncludes.ts
@@ -1,0 +1,3 @@
+export const CheckQueryIncludes = (name: string, query: string) => {
+  return name.toLowerCase().includes(query.toLowerCase());
+};
