+++ fe/src/helpers/getNumbers.ts
@@ -1,0 +1,9 @@
+export function getNumbers(from: number, to: number): number[] {
+  const numbers = [] as number[];
+
+  for (let n = from; n <= to; n += 1) {
+    numbers.push(n);
+  }
+
+  return numbers;
+}
