+++ fe/src/pages/Cart/Cart.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-
 import { CartItem } from '../../components/CartItem';
 import { CartItemCalculator } from '../../components/CartItemCalculator';
 import { Card } from '../../types/Card';
